изменено:      adb/reframe.sh 	изменено:      logo_burn.sh

+++ adb/reframe.sh
@@ -6,6 +6,11 @@ if ! [  -f "/data/data/com.termux/files/home/storage/downloads/default_wallpaper
     curl -L -o /data/data/com.termux/files/home/storage/downloads/default_wallpaper.jpg    https://github.com/definitly486/Lenovo_Tab_3_7_TB3-730X/releases/download/apk/default_wallpaper.jpg
 fi
 
+if ! [  -f "framework-res.apk_mod" ]; then
+    echo "файла framework-res.apk_mod не существует"
+    exit 
+fi
+
 su - root -c 'mount -o rw,remount /system'
 su - root -c 'mv framework-res.apk_mod /system/framework/'
 cd /system/framework
